refactor: rename providers module to provider and update references
