+++ src/market_data_ingestor/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod error;
 pub mod io;
 pub mod models;
-pub mod providers;
+pub mod provider;
